requirements.txt
pandas == 2.1.4
numpy == 1.26.2
matplotlib == 3.8.2
Scikit-Learn == 1.3.2
TensorFlow == 2.15.0
Streamlit == 1.29.0
Pilow == 10.1.0